Extracted localStorage to separate module
